Move insensitive comparisons to expressions package
